dev_requirements.txt
django-test-pep8>=0.1
Fabric>=1.6
django-devserver>=0.6
django-debug-toolbar>=0.9
